pyproject.toml
[project]
name = "ai-first-app"
version = "0.1.0"
description = "An experiment in building software where the LLM is the core, not an add-on"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
